+++ AllMyLights/Broker.cs
@@ -1,11 +1,10 @@
-﻿using AllMyLights.Connectors.Sinks;
-using AllMyLights.Connectors.Sources;
-using NLog;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reactive.Linq;
-using System.Reactive;
+using AllMyLights.Connectors.Sinks;
+using AllMyLights.Connectors.Sources;
+using NLog;
 
 namespace AllMyLights
 {