+++ AllMyLights/Common/ColorConverter.cs
@@ -71,6 +71,9 @@ namespace AllMyLights
 
         public static  OpenRGB.NET.Models.Color ToOpenRGBColor(this Color color) => new OpenRGB.NET.Models.Color(color.R, color.G, color.B);
 
+        public static string ToHexString(this Color c) => $"#{c.R:X2}{c.G:X2}{c.B:X2}";
+
+        public static string ToHexString(this OpenRGB.NET.Models.Color c) => $"#{c.R:X2}{c.G:X2}{c.B:X2}";
 
         private static void FromHex(string hex, out byte a, out byte r, out byte g, out byte b)
         {