+++ AllMyLights/Connectors/ConnectorFactory.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
 using System.Reactive.Linq;
@@ -8,10 +9,13 @@ using AllMyLights.Connectors.Sinks.OpenRGB;
 using AllMyLights.Connectors.Sinks.Wallpaper;
 using AllMyLights.Connectors.Sources;
 using AllMyLights.Connectors.Sources.Mqtt;
+using AllMyLights.Connectors.Sources.OpenRGB;
+using AllMyLights.Extensions;
 using AllMyLights.Models.OpenRGB;
 using AllMyLights.Platforms;
 using MQTTnet;
 using NLog;
+using OpenRGB.NET;
 
 namespace AllMyLights.Connectors
 {
@@ -26,11 +30,20 @@ namespace AllMyLights.Connectors
             Configuration = configuration;
         }
 
-        private OpenRGBSink GetOpenRGBSinkInstance(OpenRGBSinkOptions options) =>
-            new OpenRGBSink(options, client: new OpenRGB.NET.OpenRGBClient(
-                ip: options?.Server ?? "127.0.0.1",
-                port: options?.Port ?? 6742,
-                autoconnect: false));
+        private static Dictionary<string, IOpenRGBClient> OpenRGBClientRegistry { get; } = new Dictionary<string, IOpenRGBClient>();
+
+
+        private IOpenRGBClient GetOpenRGBClientInstance(string server, int? port)
+        {
+            var portOrDefault = port ?? 6742;
+            var serverOrDefault = server ?? "127.0.0.1";
+
+            return OpenRGBClientRegistry.SetDefault($"{serverOrDefault}:{portOrDefault}", () => new OpenRGBClient(
+                ip: serverOrDefault,
+                port: portOrDefault,
+                autoconnect: false
+            ));
+        }
 
         public ISource[] GetSources()
         {
@@ -38,6 +51,11 @@ namespace AllMyLights.Connectors
             return Configuration.Sources.Select<SourceOptions, ISource>(sourceOptions => sourceOptions switch
             {
                 MqttSourceOptions options => new MqttSource(options, new MqttFactory().CreateMqttClient()),
+                OpenRGBSourceOptions options => new OpenRGBSource(
+                    options,
+                    GetOpenRGBClientInstance(options.Server, options.Port),
+                    Observable.Interval(TimeSpan.FromMilliseconds(options.PollingInterval ?? 1000))
+                ),
                 _ => throw new NotImplementedException($"Source for type {sourceOptions.Type} not implemented")
             }).ToArray();
         }
@@ -47,7 +65,7 @@ namespace AllMyLights.Connectors
             Logger.Info($"Configuring {Configuration.Sinks.Count()} sinks");
             return Configuration.Sinks.Select<SinkOptions, ISink>(sinkOptions => sinkOptions switch
             {
-                OpenRGBSinkOptions options => GetOpenRGBSinkInstance(options),
+                OpenRGBSinkOptions options => new OpenRGBSink(options, GetOpenRGBClientInstance(options.Server, options.Port)),
                 WallpaperSinkOptions options => new WallpaperSink(options, Desktop.GetPlatformInstance()),
                 ChromaSinkOptions options => new ChromaSink(options, new ChromaClient(HttpClient), Observable.Interval(TimeSpan.FromSeconds(2))),
                 _ => throw new NotImplementedException($"Sinks for type {sinkOptions.Type} not implemented")