+++ AllMyLights/Connectors/Sources/OpenRGB/DeviceState.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections.Generic;
+
+namespace AllMyLights.Connectors.Sources.OpenRGB
+{
+    public struct DeviceState
+    {
+        public IEnumerable<string> Colors;
+
+        public DeviceState(IEnumerable<string> colors)
+        {
+            Colors = colors;
+        }
+    }
+}