+++ AllMyLights/Connectors/Sources/OpenRGB/OpenRGBSourceOptions.cs
@@ -1,0 +1,14 @@
+﻿
+namespace AllMyLights.Connectors.Sources.OpenRGB
+{
+    public class OpenRGBSourceOptions: SourceOptions
+    {
+
+        public string Server { get; set; }
+        public int? Port { get; set; }
+
+        public int? PollingInterval { get; set; }
+    }
+}
+
+