+++ AllMyLights/Extensions/Dictionaries.cs
@@ -12,5 +12,20 @@ namespace AllMyLights.Extensions
             dict.TryGetValue(key, out TValue value);
             return value ?? defaultValue;
         }
+
+        public static TValue SetDefault<TKey, TValue>(
+            this IDictionary<TKey, TValue> dict,
+            TKey key,
+            Func<TValue> produceValue)
+        {
+            if (!dict.TryGetValue(key, out TValue value))
+            {
+                var added = produceValue();
+                dict.Add(key, added);
+                return added;
+            }
+
+            return value;
+        }
     }
 }