+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [v7]
+
+### Sources
+- Add `OpenRGB` source to enable use cases in which OpenRGB is the source of truth
 ## [v6]
 
 ### Sinks
@@ -56,6 +60,7 @@ via JsonPath or color conversions reusable across different source and sink type
 - Add option to override the channel layout of a device
 - Add option to override the channel layout of a zone
 
+[v7]: https://github.com/sparten11740/allmylights/compare/v6...v7
 [v6]: https://github.com/sparten11740/allmylights/compare/v5...v6
 [v5]: https://github.com/sparten11740/allmylights/compare/v4...v5
 [v4]: https://github.com/sparten11740/allmylights/compare/v3...v4